+++ packages/core-react/CHANGELOG.md
@@ -1,10 +1,37 @@
 # @livepeer/core-react
 
+## 3.2.7
+
+### Patch Changes
+
+- [#577](https://github.com/livepeer/ui-kit/pull/577) [`3b41b4d`](https://github.com/livepeer/ui-kit/commit/3b41b4d30c9ce7dfade6e6360bcc464421e17af0) Thanks [@0xcadams](https://github.com/0xcadams)! - **Fix:** added `oldest_buffer_timestamp` to heartbeat events.
+
+- Updated dependencies [[`3b41b4d`](https://github.com/livepeer/ui-kit/commit/3b41b4d30c9ce7dfade6e6360bcc464421e17af0)]: 
+  - @livepeer/core@3.2.7
+
+## 3.2.6
+
+### Patch Changes
+
+- [#575](https://github.com/livepeer/ui-kit/pull/575) [`a3a5034`](https://github.com/livepeer/ui-kit/commit/a3a503497803f54fddef0fd0dffee96872944ffa) Thanks [@0xcadams](https://github.com/0xcadams)! - **Fix:** updated all dependencies and changed to have a single heartbeat event when multiple heartbeats are attempted to be sent.
+
+- Updated dependencies [[`a3a5034`](https://github.com/livepeer/ui-kit/commit/a3a503497803f54fddef0fd0dffee96872944ffa)]:
+  - @livepeer/core@3.2.6
+
+## 3.2.5
+
+### Patch Changes
+
+- [#573](https://github.com/livepeer/ui-kit/pull/573) [`2a1b4e4`](https://github.com/livepeer/ui-kit/commit/2a1b4e499fe829face2606d2ae2a23443db2ff1e) Thanks [@0xcadams](https://github.com/0xcadams)! - **Fix:** moved to using `nanoid` for generating secure IDs and added IDs to all metrics.
+
+- Updated dependencies [[`2a1b4e4`](https://github.com/livepeer/ui-kit/commit/2a1b4e499fe829face2606d2ae2a23443db2ff1e)]:
+  - @livepeer/core@3.2.5
+
 ## 3.2.4
 
 ### Patch Changes
 
-- [#571](https://github.com/livepeer/ui-kit/pull/571) [`96ab02b`](https://github.com/livepeer/ui-kit/commit/96ab02b18b4cd68aa25b2c589e78a42dfa8c8ff1) Thanks [@0xcadams](https://github.com/0xcadams)! - **Fix:** fixed logic for falling back to HLS. 
+- [#571](https://github.com/livepeer/ui-kit/pull/571) [`96ab02b`](https://github.com/livepeer/ui-kit/commit/96ab02b18b4cd68aa25b2c589e78a42dfa8c8ff1) Thanks [@0xcadams](https://github.com/0xcadams)! - **Fix:** fixed logic for falling back to HLS.
 
 - Updated dependencies [[`96ab02b`](https://github.com/livepeer/ui-kit/commit/96ab02b18b4cd68aa25b2c589e78a42dfa8c8ff1)]:
   - @livepeer/core@3.2.4